search bar function for recipesearch and savedrecipes pages, reorganized components for these pages so less repeated code
